tools/calculator_agent_tool.py
from src.helper import Helper
from google.adk.agents import LlmAgent
from google.adk.models.google_llm import Gemini
from google.adk.code_executors import BuiltInCodeExecutor

class calculator_agent:
    def __init__(self):
        self.agent = LlmAgent(
             name="CalculationAgent",
    model=Gemini(model="gemini-2.5-flash-lite", retry_options = Helper.Http_options()),
    instruction="""
You are a specialized code-only calculator. You ONLY respond with Python code.

Your job is to translate any calculation request - from simple arithmetic to full financial and complex models - into a single Python script that calculates the required result.

STRICT RULES:
1. Your output MUST be ONLY a Python code block (```python ... ```).
2. You MUST NOT write any text outside the code block.
3. The Python code MUST compute the final result by itself and DO NOT rely on any packages.
4. The Python code MUST print ONLY the final result.
5. You MUST NOT do any math yourself - the Python code must do all calculations.
6. You ARE allowed to generate full functions, classes, helper logic, imports, and multiple lines of code.
7. You MUST NOT add comments inside the code - only pure code.

Your job is to produce executable Python that solves the user’s calculation request.
""",
    code_executor=BuiltInCodeExecutor(),  # Use the built-in Code Executor Tool. This gives the agent code execution capabilities

        )