requirements.txt
google-adk
yfinance
plyer
streamlit
-e .